% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-matrix.R,
%   R/aggregation-matrix-default.R, R/aggregation-matrix-membership-difference.R
\docType{methods}
\name{Aggregate}
\alias{Aggregate}
\alias{Aggregate,Aggregation.matrix.default,character,character,logical,numeric-method}
\alias{Aggregate,Aggregation.matrix.membership.difference,character,character,logical,numeric-method}
\title{Aggregate}
\usage{
Aggregate(aggregation.matrix, factor.evaluation, resource.evaluation,
  factor.is.specific, nrfactors)


  \S4method{Aggregate}{Aggregation.matrix.default,character,character,logical,numeric}(aggregation.matrix,
  factor.evaluation, resource.evaluation, factor.is.specific, nrfactors)


  \S4method{Aggregate}{Aggregation.matrix.membership.difference,character,character,logical,numeric}(aggregation.matrix,
  factor.evaluation, resource.evaluation, factor.is.specific, nrfactors)
}
\arguments{
\item{aggregation.matrix}{aggregation.matrix}

\item{factor.evaluation}{character factor evaluation from project}

\item{resource.evaluation}{character factor evaluation from option}

\item{factor.is.specific}{logic indicates that this factor is specific for
the project}

\item{nrfactors}{numeric number of factors evaluated for each project/option}
}
\value{
numeric indicate the result factor per option. If a specific factor
is not achieved it returns -1
}
\description{
S4 method do not validate entries, since it is not exported and the data
is validated by the constructors. The validation here would be resource
consuming.
}
