% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-matrix.R,
%   R/aggregation-matrix-default.R
\docType{methods}
\name{AggregateMatrix}
\alias{AggregateMatrix}
\alias{AggregateMatrix,Aggregation.matrix.default,data.frame,data.frame,data.frame-method}
\title{AggregateMatrix}
\usage{
AggregateMatrix(aggregation.matrix, project.portfolio.as.data.frame,
  project.portfolio.specifics.as.data.frame, option.portfolio.as.data.frame)


  \S4method{AggregateMatrix}{Aggregation.matrix.default,data.frame,data.frame,data.frame}(aggregation.matrix,
  project.portfolio.as.data.frame, project.portfolio.specifics.as.data.frame,
  option.portfolio.as.data.frame)
}
\value{
data.frame

data.frame
}
\description{
S4 method to perform Aggregation.Matrix inheirited objects. The method to be
executed is selected by the Ag
# TODO(Pessoa) explicar aqui
}
