% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-matrix.R
\docType{class}
\name{Aggregation.matrix-class}
\alias{Aggregation.matrix-class}
\title{Aggregation.matrix S4 Class}
\description{
This class was included to act as an abstract class to be inherited by
concrete classes that implement their matrix in constructors.
}
\section{Slots}{

\describe{
\item{\code{name}}{character}
}}

