% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-matrix-default.R
\docType{class}
\name{Aggregation.matrix.default-class}
\alias{Aggregation.matrix.default-class}
\title{Aggregation.matrix.default}
\description{
This class represents the default aggregation matrix. This matrix was defined
by ... # TODO(Cosenza)
}
