% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-matrix-membership-difference.R
\docType{class}
\name{Aggregation.matrix.membership.difference-class}
\alias{Aggregation.matrix.membership.difference-class}
\title{Aggregation.matrix.membership.difference}
\description{
This class represents extends the Aggregation.matrix S4 and implements  the
Membership Difference aggregation matrix.
}
\seealso{
Aggregation.matrix
}
