% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-matrix-default.R
\name{Agregate}
\alias{Agregate}
\title{Agregate}
\usage{
Agregate(factor.evaluation, resource.evaluation, factor.is.specific, nrfactors)
}
\arguments{
\item{factor.evaluation}{character factor evaluation from project}

\item{resource.evaluation}{character factor evaluation from option}

\item{factor.is.specific}{logic indicates that this factor is specific for
the project}

\item{nrfactors}{numeric number of factors evaluated for each project/option}
}
\value{
numeric indicate the result factor per option. If a specific factor
is not achived it returns -1
}
\description{
This function do not validate entries, since it is not exported and the data
is validated by the constructors. The validation here would be resource
consuming.
}
