% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-portfolio.R, R/project-portfolio.R
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,Option.portfolio-method}
\alias{as.data.frame,Project.portfolio-method}
\title{as.data.frame}
\usage{
as.data.frame(x, ...)

\S4method{as.data.frame}{Option.portfolio}(x)

\S4method{as.data.frame}{Project.portfolio}(x, get.specific = FALSE)
}
\arguments{
\item{x}{Option.portfolio S4 object}

\item{(x, }{get.specific = FALSE)
\itemize{
\item x is a Project.portfolio S4 objec
\item get.specific indicates if the return is a data.frame with factor
evaluations or with the information about which factors are specific to
a project.
}}
}
\value{
data.frame

data.frame
}
\description{
Generic S4 method to \code{\link{as.data.frame}}.
}
\examples{
\dontrun{as.data.frame(option.portfolio)}

\dontrun{as.data.frame(option.portfolio, TRUE)}


}
