% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppeCosenza.R
\docType{class}
\name{Coppe.cosenza-class}
\alias{Coppe.cosenza-class}
\title{Coppe.cosenza S4 Class}
\description{
Coppe.cosenza S4 class represents the solution of the COPPE-Cosenza method.
# TODO(Pessoa)Ampliar aqui.
}
\section{Slots}{

\describe{
\item{\code{result}}{data.frame}

\item{\code{messages}}{list}
}}

