% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppe-cosenza.R
\docType{methods}
\name{Coppe.cosenza}
\alias{Coppe.cosenza}
\alias{Coppe.cosenza,ANY,ANY,ANY,ANY-method}
\alias{Coppe.cosenza,Project.portfolio,Option.portfolio,Factors.under.consideration,missing-method}
\alias{Coppe.cosenza,Project.portfolio,Option.portfolio,Factors.under.consideration,character-method}
\title{Coppe.cosenza}
\usage{
Coppe.cosenza(x, y, z, k)

\S4method{Coppe.cosenza}{ANY,ANY,ANY,ANY}(x)


  \S4method{Coppe.cosenza}{Project.portfolio,Option.portfolio,Factors.under.consideration,missing}(x,
  y, z, k)


  \S4method{Coppe.cosenza}{Project.portfolio,Option.portfolio,Factors.under.consideration,character}(x,
  y, z, k = "default")
}
\arguments{
\item{Arguments}{(ANY) A call to \code{Coppe.cosenza( )} with no parameters
will return an error message for missing argument.}

\item{Arguments}{\itemize{
\item Project.portfolio S4 object
\item Option.portfolio S4 object
\item Factors.under.consideration S4 object
\item character - the name of Aggregation.matrix to be used. If not provided
the default implementation will be used}}
}
\value{
Coppe.cosenza S4 object
}
\description{
S4 method to construc Coppe.cosenza objects. It solves de
# TODO(Pessoa) explicar aqui
}
