% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppeCosenza.R
\docType{package}
\name{coppeCosenza-package}
\alias{coppeCosenza}
\alias{coppeCosenza-package}
\title{COPPE-Cosenza Fuzzy Hierarchy Model (coppeCosenza)}
\description{
This package .......
}
\details{
The package also includes several facilities for ......

Currently, coppeCosenza provides implementation to support .....:

\itemize{
\item Stratified.ceg.model
\item Stratified.staged.tree
\item Stratified.event.tree
}

These classes are implemented as S4 classes and have constructor methods with
the same name as the class. A \code{plot} method is also provided.
}

