% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppeCosenza-method.R
\name{CoppeCosenzaMethod}
\alias{CoppeCosenzaMethod}
\title{CoppeCosenzaMethod}
\usage{
CoppeCosenzaMethod(project.portfolio, option.portfolio,
  factors.under.consideration, ...)
}
\description{
CoppeCosenzaMethod
}
