% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppeCosenzaR.R
\docType{package}
\name{coppeCosenzaR-package}
\alias{coppeCosenzaR}
\alias{coppeCosenzaR-package}
\title{COPPE-Cosenza Fuzzy Hierarchy Model (coppeCosenzaR)}
\description{
This package .......
}
\details{
accepted.availability <- c("Excellent", "Good", "Regular", "Weak", "Void", "Zero", "Inexistent")
The package also includes several facilities for ......

Currently, coppeCosenza provides implementation to support .....:

\itemize{
\item Stratified.ceg.model
\item Stratified.staged.tree
\item Stratified.event.tree
}

These classes are implemented as S4 classes and have constructor methods with
the same name as the class. A \code{plot} method is also provided.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{XXX@XXX.xxx}
}

}
\author{
\strong{Maintainer}: Leonardo Pessôa \email{XXX@XXX.xxx}

Authors:
\itemize{
  \item Pier Taranti \email{ptaranti@gmail.com}
}

}
