% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{Factor}
\alias{Factor}
\title{Factor Constructor}
\usage{
Factor(name)
}
\arguments{
\item{name}{character the factor name. If a diferent type is provided, it is
casted to character using \code{\link{as.charactere()}} function.}
}
\value{
a \code{\link{Factor}} S4 object
}
\description{
Factor(name) is a cosntructor to Factor S4 objects.
Factor S4 class contains a sigle slot with the factor name.
}
\examples{
factor <- Factor("name")
Factor("name")


}
