% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors-of-interest.R
\docType{class}
\name{Factors.of.interest-class}
\alias{Factors.of.interest-class}
\title{Factors.of.interest S4 Class}
\description{
Factors.of.interest S4 class contains a list of S4 Factor objects.
This list is used as parameter when construction the output from Coppe-Cosenza
 method.
}
\section{Slots}{

\describe{
\item{\code{list.of.factors}}{list of Factor.
Has one or more distinct S4 Factor objects.}
}}

