% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors-of-interest.R
\name{Factors.of.interest}
\alias{Factors.of.interest}
\title{Factors.of.interest Constructor}
\usage{
Factors.of.interest(list.of.factors)
}
\arguments{
\item{list.of.factors}{list of Factor S4 objects}
}
\value{
a \code{\link{Factors.of.interest}} S4 object
}
\description{
Factors.of.interest is a constructor. Factor elements inserted in
list.of.factors are type-checked as S4 coppeCosenza::Factor objects. They
must have distinct names.
}
\examples{
Factors.of.interest(list(Factor("factor1"), Factor("factor2"),
Factor("factor3")))

}
