% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors-under-consideration.R
\docType{class}
\name{Factors.under.consideration-class}
\alias{Factors.under.consideration-class}
\title{Factors.under.consideration S4 Class}
\description{
Factors.under.consideration S4 class contains a list of S4 Factor objects.
This list is   used to construct Projec objects.
}
\section{Slots}{

\describe{
\item{\code{list.of.factors}}{list of Factor}
}}

