% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_under_consideration.R
\name{Factors.under.consideration}
\alias{Factors.under.consideration}
\title{Factors.under.consideration(list.of.factors)}
\usage{
Factors.under.consideration(list.of.factors)
}
\arguments{
\item{list.of.factors}{list of Factor S4 objects}
}
\value{
a \code{\link{Factors.under.consideration}} S4 object
}
\description{
Factors.under.consideration(list.of.factors) is a constructor.
}
\examples{
Factors.under.consideration(list(Factor("factor1"), Factor("factor2"), Factor("factor3")))

}
