% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors-of-interest.R
\name{getFactorsOfInterestNames}
\alias{getFactorsOfInterestNames}
\title{getFactorsOfInterestNames}
\usage{
getFactorsOfInterestNames(factors.of.interest)
}
\arguments{
\item{factors.of.interest}{S4 Factors.of.interest object}
}
\value{
vector of character
}
\description{
It provides a sorted vector with the names of factors.
}
\examples{
\dontrun{getFactorsOfInterestNames(factors.of.interest)}

}
