% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors-under-consideration.R
\name{getFactorsUnderConsiderationNames}
\alias{getFactorsUnderConsiderationNames}
\title{getFactorsUnderConsiderationNames}
\usage{
getFactorsUnderConsiderationNames(factors.under.consideration)
}
\arguments{
\item{factors.under.consideration}{}
}
\value{
vector of character
}
\description{
It provides a sorted vector with the names of factors.
}
\examples{
\dontrun{getFactorsUnderConsiderationNames(factors.under.consideration)}

}
