% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{getOptionFactorsNames}
\alias{getOptionFactorsNames}
\title{getOptionFactorsNames}
\usage{
getOptionFactorsNames(option)
}
\arguments{
\item{option}{an Option S4 object}
}
\value{
It provides a sorted vector with the names of factors in an option.
}
\description{
This function returns a sorted vector with all the factors names in a Option
S4 object
}
\examples{
\dontrun{getOptionFactorsNames(option)}

}
