% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-portfolio.R
\name{getOptionPortfolioAsDataFrame}
\alias{getOptionPortfolioAsDataFrame}
\title{Title}
\usage{
getOptionPortfolioAsDataFrame(option.portfolio)
}
\arguments{
\item{option.portfolio}{}
}
\description{
Title
}

