% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-portfolio.R
\name{getOptionPortfolioFactors}
\alias{getOptionPortfolioFactors}
\title{getOptionPortfolioFactors}
\usage{
getOptionPortfolioFactors(option.portfolio)
}
\arguments{
\item{option.portfolio}{S4 Option.portfolio object}
}
\value{
list of Factor S4 objects
}
\description{
function that provides a list of Factor S4 objects presents in a
Option.portfolio S4 object
}
\examples{
\dontrun{getOptionPortfolioFactors(option.portfolio)}

}
