% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-portfolio.R
\name{getOptionPortfolioFactorsAsDataFrame}
\alias{getOptionPortfolioFactorsAsDataFrame}
\title{Title}
\usage{
getOptionPortfolioFactorsAsDataFrame(option.portfolio)
}
\arguments{
\item{option.portfolio}{}
}
\description{
Title
}
