% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{getProjectFactorsNames}
\alias{getProjectFactorsNames}
\title{getProjectFactorsNames}
\usage{
getProjectFactorsNames(project)
}
\arguments{
\item{project}{an Project S4 object}
}
\value{
It provides a sorted vector with the names of factors in an project
}
\description{
This function returns a sorted vector with all the factors names in a Project
S4 object
}
\examples{
\dontrun{getProjectFactorsNames(project)}

}
