% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{getProjectFactorsSpecific}
\alias{getProjectFactorsSpecific}
\title{getProjectFactorsSpecific}
\usage{
getProjectFactorsSpecific(project)
}
\arguments{
\item{project}{an Project S4 object}
}
\value{
It provides a sorted vector with the names of factors in an project
which were classified as specific to the project under discussion.
}
\description{
This function returns a sorted vector with all the factors names in a Project
S4 object which were classified as specific to the project under discussion.
}
\examples{
\dontrun{getProjectFactorsSpecific(project)}

}
