% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-portfolio.R
\name{getProjectPortfolioAsDataFrame}
\alias{getProjectPortfolioAsDataFrame}
\title{Title}
\usage{
getProjectPortfolioAsDataFrame(project.portfolio)
}
\arguments{
\item{project.portfolio}{}
}
\description{
Title
}

