% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-portfolio.R
\name{getProjectPortfolioFactors}
\alias{getProjectPortfolioFactors}
\title{getProjectPortfolioFactors}
\usage{
getProjectPortfolioFactors(project.portfolio)
}
\arguments{
\item{project.portfolio}{}
}
\value{
vector of character
}
\description{
function that provides a sorted vector with factors from the project list.
}
\examples{
\dontrun{getProjectPortfolioFactors(project.portfolio)}

}
