% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-portfolio.R
\name{getProjectPortfolioFactorsAsDataFrame}
\alias{getProjectPortfolioFactorsAsDataFrame}
\title{Title}
\usage{
getProjectPortfolioFactorsAsDataFrame(project.portfolio)
}
\arguments{
\item{project.portfolio}{}
}
\description{
Title
}
