% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-portfolio.R
\name{getProjectPortfolioNames}
\alias{getProjectPortfolioNames}
\title{getProjectPortfolioNames}
\usage{
getProjectPortfolioNames(project.portfolio)
}
\arguments{
\item{project.portfolio}{S4 Project.portfolio object}
}
\value{
vector of character
}
\description{
function that provides a sorted vector with project names.
}
\examples{
\dontrun{getProjectPortfolioNames(project.portfolio)}

}
