% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-portfolio.R
\name{getProjectPortfolioSpecificsAsDataFrame}
\alias{getProjectPortfolioSpecificsAsDataFrame}
\title{Title}
\usage{
getProjectPortfolioSpecificsAsDataFrame(project.portfolio)
}
\arguments{
\item{project.portfolio}{}
}
\description{
Title
}

