% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs.R
\name{LoadAgregationMatrix}
\alias{LoadAgregationMatrix}
\title{LoadAgregationMatrix}
\usage{
LoadAgregationMatrix(x)
}
\arguments{
\item{x}{address to a file with AgregationMatrix}
}
\description{
LoadAgregationMatrix
}
