% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\docType{class}
\name{Option-class}
\alias{Option-class}
\title{Option S4 Class}
\description{
Option S4 class.  ... TODO(Pessoa) VRF eAmpliar
}
\section{Slots}{

\describe{
\item{\code{name}}{character}

\item{\code{option.resources}}{Option.resources}
}}

