% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-factor-availability.R
\docType{class}
\name{Option.factor.availability-class}
\alias{Option.factor.availability-class}
\title{Option.factor.availability S4 Class}
\description{
Option.factor.availability S4 class. It defines the availability to be used in
association to a factor when evaluatig projects ... TODO(Pessoa) VRF eAmpliar
}
\section{Slots}{

\describe{
\item{\code{factor}}{Factor S4 class}

\item{\code{availability}}{char, must mach the scale of degrees to be used}
}}

