% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-portfolio.R
\docType{class}
\name{Option.portfolio-class}
\alias{Option.portfolio-class}
\title{Option.portfolio S4 Class}
\description{
Option.portfolio S4 class contains a list of S4 Option objects.
# TODO (Pessoa) explicar que é input com o resources.portfolio e a matrix de
agregacao.
}
\section{Slots}{

\describe{
\item{\code{list.of.option}}{list of Option S4 objects}
}}

