% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-portfolio.R
\name{Option.portfolio}
\alias{Option.portfolio}
\title{Option.portfolio Constructor}
\usage{
Option.portfolio(list.of.option)
}
\arguments{
\item{list.of.option}{list of Option S4 objects}
}
\value{
a \code{\link{Option.portfolio}} S4 object
}
\description{
Option.portfolio(list.of.option) is a constructor to Option.portfolio S4 objects.
}
\examples{
\dontrun{Option.portfolio(list(project1, project2, project3))}
# TODO(Taranti) inserir exemplo

}
