% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{Option}
\alias{Option}
\title{Option Constructor function}
\usage{
Option(name, option.resources)
}
\arguments{
\item{name}{charactere}

\item{option.resources}{Option.criteria S4 object}
}
\value{
a \code{\link{Option}} S4 object
}
\description{
Constructs a Option S4 object.  ... TODO(Pessoa) VRF e Ampliar
}
\examples{
\dontrun{Option <- Option(name, option.resources)}



}

