% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-resources.R
\docType{class}
\name{Option.resources-class}
\alias{Option.resources-class}
\title{Option.resources S4 Class}
\description{
Option.resources S4 class contains a list of one or more S4
Option.factor.availability objects. This list is type-checked and used to
construct Option objects.
}
\section{Slots}{

\describe{
\item{\code{list.of.factor.availability}}{list of Option.factor.availability}
}}

