% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-resources.R
\name{Option.resources}
\alias{Option.resources}
\title{Option.resources Constructor}
\usage{
Option.resources(list.of.factor.availability)
}
\arguments{
\item{list}{list of Option.factor.availability S4 objects}
}
\value{
a \code{\link{Option.resources}} S4 object
}
\description{
Option.resources(list) is a constructor to Factor S4 objects.
}
\examples{

# TODO(Taranti) inserir exemplo

}

