% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{class}
\name{Project-class}
\alias{Project-class}
\title{Project S4 Class}
\description{
Project S4 class.  ... TODO(Pessoa) VRF eAmpliar
}
\section{Slots}{

\describe{
\item{\code{name}}{character}

\item{\code{project.criteria}}{Project.criteria}
}}

