% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-criteria.R
\docType{class}
\name{Project.criteria-class}
\alias{Project.criteria-class}
\title{Project.criteria S4 Class}
\description{
Project.criteria S4 class contains a list of S4 Project.criterion objects.
This list is   used to construct Projec objects, and is type checked.
}
\section{Slots}{

\describe{
\item{\code{list.of.project.criterion}}{list of Project.criterion}
}}

