% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-criteria.R
\name{Project.criteria}
\alias{Project.criteria}
\title{Project.criteria Constructor}
\usage{
Project.criteria(list.of.project.criterion)
}
\arguments{
\item{list}{list of Project.criterion S4 objects}
}
\value{
a \code{\link{Project.criteria}} S4 object
}
\description{
Project.criteria(list) is a constructor to Factor S4 objects.
}
\examples{
# TODO(Taranti) inserir exemplo

}

