% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-criterion.R
\docType{class}
\name{Project.criterion-class}
\alias{Project.criterion-class}
\title{Project.criterion S4 Class}
\description{
Project.criterion S4 class. It defines the criterion to be used in
association to a factor when evaluating projects.
}
\details{
The accepted degrees are: "Cr", "C", "LC","I"
}
\section{Slots}{

\describe{
\item{\code{factor}}{Factor S4 class}

\item{\code{importance.degree}}{character, must mach the scale of degrees to be used}

\item{\code{specific}}{logical indicates the considered factors is specific for the
project under consideration}
}}

