% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-criterion.R
\name{Project.criterion}
\alias{Project.criterion}
\title{Project.criterion}
\usage{
Project.criterion(factor, importance.degree, specific)
}
\arguments{
\item{factor}{Factor S4 class}

\item{importance.degree}{character, must mach one item of the scale of
degrees to be used ("Cr", "C", "LC","I")}

\item{specific}{logical indicates the considered factors is specific for the
project under consideration#'}
}
\value{
a \code{\link{Project.criterion}} S4 object
}
\description{
This function is a constructor to Project.criterion S4 class. It defines the
criterion to be used in association to a factor when evaluating projects.
}
\examples{
\dontrun{Project.criterion <- Project.criterion(factor, importance.degree, specific)}
Project.criterion(Factor("fator1"), "LC", FALSE)

}
