% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-portfolio.R
\docType{class}
\name{Project.portfolio-class}
\alias{Project.portfolio-class}
\title{Project.portfoio S4 Class}
\description{
Project.portfoio S4 class contains a list of S4 Project objects.
# TODO (Pessoa) explicar que é input com o resources.portfolio e a matrix de
agregacao.
}
\section{Slots}{

\describe{
\item{\code{list.of.project}}{list of Project S4 objects}
}}

