% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-portfolio.R
\docType{methods}
\name{Project.portfolio}
\alias{Project.portfolio}
\alias{Project.portfolio,ANY,ANY-method}
\alias{Project.portfolio,list,ANY-method}
\alias{Project.portfolio,data.frame,data.frame-method}
\title{Project.portfolio Constructor}
\usage{
Project.portfolio(x, y, ...)

\S4method{Project.portfolio}{ANY,ANY}(x, y, ...)

\S4method{Project.portfolio}{list,ANY}(x)

\S4method{Project.portfolio}{data.frame,data.frame}(x, y)
}
\arguments{
\item{list.of.project}{list of Project S4 objects}

\item{Arguments}{(ANY) \cr
A call to \code{Project.portfolio( )} with no parameters will return
an error message for missing argument.}
}
\value{
a \code{\link{Project.portfolio}} S4 object

S4 class
}
\description{
Project.portfolio(list.of.project) is a constructor to Project.portfolio S4 objects.
}
\examples{
\dontrun{Project.portfolio(list(project1, project2, project3))}
# TODO(Taranti) inserir exemplo

Project.portfolio Constructor

}
