% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-portfolio.R
\name{Project.portfolio2}
\alias{Project.portfolio2}
\title{Title}
\usage{
Project.portfolio2(project.portfolio.as.data.frame,
  project.portfolio.specifics.as.data.frame)
}
\arguments{
\item{project.portfolio.specifics.as.data.frame}{}
}
\description{
Title
}
