% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{Project}
\alias{Project}
\title{Project Constructor function}
\usage{
Project(name, project.criteria)
}
\arguments{
\item{name}{character}

\item{project.criteria}{Project.criteria S4 object}
}
\value{
a \code{\link{Project}} S4 object
}
\description{
Constructs a Project S4 object.  ... TODO(Pessoa) VRF e Ampliar
}
\examples{
\dontrun{Project <- Project(name, project.criteria)}



}
