% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-matrix.R,
%   R/aggregation-matrix-default.R,
%   R/aggregation-matrix-membership-difference.R, R/factor.R,
%   R/factors-of-interest.R, R/option-factor-availability.R,
%   R/option-resources.R, R/option.R, R/option-portfolio.R,
%   R/project-criterion.R, R/project-criteria.R, R/project.R,
%   R/project-portfolio.R, R/coppe-cosenza.R
\docType{methods}
\name{show,Aggregation.matrix-method}
\alias{show,Aggregation.matrix-method}
\alias{show}
\alias{show,Aggregation.matrix.default-method}
\alias{show,Aggregation.matrix.membership.difference-method}
\alias{show,Factor-method}
\alias{show,Factors.of.interest-method}
\alias{show,Option.factor.availability-method}
\alias{show,Option.resources-method}
\alias{show,Option-method}
\alias{show,Option.portfolio-method}
\alias{show,Project.criterion-method}
\alias{show,Project.criteria-method}
\alias{show,Project-method}
\alias{show,Project.portfolio-method}
\alias{show,Coppe.cosenza-method}
\title{show}
\usage{
\S4method{show}{Aggregation.matrix}(object)

show(object)

\S4method{show}{Aggregation.matrix.default}(object)

\S4method{show}{Aggregation.matrix.membership.difference}(object)

\S4method{show}{Factor}(object)

\S4method{show}{Factors.of.interest}(object)

\S4method{show}{Option.factor.availability}(object)

\S4method{show}{Option.resources}(object)

\S4method{show}{Option}(object)

\S4method{show}{Option.portfolio}(object)

\S4method{show}{Project.criterion}(object)

\S4method{show}{Project.criteria}(object)

\S4method{show}{Project}(object)

\S4method{show}{Project.portfolio}(object)

\S4method{show}{Coppe.cosenza}(object)
}
\arguments{
\item{object}{Aggregation.matrix}

\item{Aggregation.matrix.default}{Aggregation.matrix.default}

\item{Aggregation.matrix.membership.difference}{Aggregation.matrix.membership.difference}

\item{Factor}{Factor}

\item{Factors.of.interest}{Factors.of.interest}

\item{Option.factor.availability}{Option.factor.availability}

\item{Option.resources}{Option.resources}

\item{Option}{Option}

\item{Option.portfolio}{Option.portfolio}

\item{Project.criterion}{Project.criterion}

\item{Project.criteria}{Project.criteria}

\item{Project}{Project}

\item{Project.portfolio}{Project.portfolio}

\item{Coppe.cosenza}{Coppe.cosenza}
}
\description{
show
}
