% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppe-cosenza.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,Coppe.cosenza-method}
\title{summary}
\usage{
summary(object, ...)

\S4method{summary}{Coppe.cosenza}(object)
}
\arguments{
\item{object}{Coppe.cosenza}

\item{...}{not used.}
}
\value{
summary
}
\description{
Generic S4 method to \code{\link{summary}}.
}
