% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation-matrix.R
\docType{methods}
\name{AggregateMatrix}
\alias{AggregateMatrix}
\alias{AggregateMatrix,Aggregation.matrix,data.frame,data.frame,data.frame-method}
\title{AggregateMatrix}
\usage{
AggregateMatrix(aggregation.matrix, project.portfolio.as.data.frame,
  project.portfolio.specifics.as.data.frame, option.portfolio.as.data.frame)


  \S4method{AggregateMatrix}{Aggregation.matrix,data.frame,data.frame,data.frame}(aggregation.matrix,
  project.portfolio.as.data.frame, project.portfolio.specifics.as.data.frame,
  option.portfolio.as.data.frame)
}
\arguments{
\item{aggregation.matrix}{aggregation.matrix}

\item{project.portfolio.as.data.frame}{project.portfolio.as.data.frame}

\item{project.portfolio.specifics.as.data.frame}{project.portfolio.specifics.as.data.frame}

\item{option.portfolio.as.data.frame}{option.portfolio.as.data.frame}
}
\value{
data.frame

data.frame
}
\description{
S4 method to perform Aggregation.Matrix inheirited objects.
If a implementation is not provided to a specific aggregation matrix, this
implementation will be used.
So it allows using different aggregation matrices. Such feature provides
flexibility regarding compensatory effects among the criteria.  Therefore it
can be adjusted to different multicriteria problems. It is important to
highlight that the entries must be compliant to the original described
categories,  using the same linguistic variables present in the default
aggregation matrix.
}
