% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-portfolio.R, R/project-portfolio.R
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,Option.portfolio-method}
\alias{as.data.frame,Project.portfolio-method}
\title{as.data.frame}
\usage{
as.data.frame(x, row.names, optional, ...)

\S4method{as.data.frame}{Option.portfolio}(x)

\S4method{as.data.frame}{Project.portfolio}(x, row.names = NA,
  optional = FALSE)
}
\arguments{
\item{x}{Option.portfolio or Project.portfolio}

\item{row.names}{not used. It is inherited from \code{base::as.data.frame()}}

\item{optional}{logical. To be used with  Project.portfolio. Indicates if the
return is a data.frame with factor evaluations or with the information about
 which factors are specific to a project.
 The default is \code{optional = FALSE}}

\item{...}{not used.}
}
\value{
data.frame
}
\description{
This S4 method masks the \code{base::as.data.frame()} S3 function. If a call
uses parameters other then the expected by this package, then it will be
forward to the S3 function.
}
\examples{
\dontrun{as.data.frame(option.portfolio)}

\dontrun{as.data.frame(project.portfolio, option = TRUE)}
\dontrun{as.data.frame(project.portfolio,  , TRUE)}
\dontrun{as.data.frame(project.portfolio, ANY, FALSE)}
\dontrun{as.data.frame(project.portfolio, option = FALSE)}
\dontrun{as.data.frame(project.portfolio)} This infer option is FALSE, too.

}
