% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppe-cosenza.R
\docType{class}
\name{Coppe.cosenza-class}
\alias{Coppe.cosenza-class}
\title{Coppe.cosenza S4 Class}
\description{
Coppe.cosenza S4 class represents the solution of the COPPE-Cosenza method.
In order to do so, this S4 class contains the final evaluation of the options
regarding the studied projects. It presents a data frame presenting the
final evaluation of the options regarding each project.
If an option does not satisfies project´s specific factors, the option is
discarded (a veto operation), with the value NA.  The result also
presents relevant  messages list, describing if some evaluation could not be
performed due to entry failures or missing evaluations.
}
\section{Slots}{

\describe{
\item{\code{result}}{data.frame}

\item{\code{projects.names}}{character}

\item{\code{options.names}}{character}

\item{\code{factors.of.interest}}{Factors.of.interest}

\item{\code{aggregation.matrix}}{Aggregation.matrix}

\item{\code{messages}}{character}
}}

