% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppeCosenza.R
\docType{methods}
\name{Coppe.cosenza}
\alias{Coppe.cosenza}
\alias{Coppe.cosenza,ANY,ANY,ANY-method}
\alias{Coppe.cosenza,Project.portfolio,Option.portfolio,Factors.under.consideration-method}
\title{Coppe.cosenza}
\usage{
Coppe.cosenza(x, y, z, k, ...)

\S4method{Coppe.cosenza}{ANY,ANY,ANY}(x, y, z, k, ...)


  \S4method{Coppe.cosenza}{Project.portfolio,Option.portfolio,Factors.under.consideration}(x,
  y, z)
}
\arguments{
\item{Arguments}{(ANY) A call to \code{Coppe.cosenza( )} with no parameters
will return an error message for missing argument.}

\item{Arguments}{\itemize{
\item x  a Project.portfolio S4 object
\item Option.portfolio S4 object
\item Factors.under.consideration S4 object}}
}
\value{
Coppe.cosenza S4 object
}
\description{
S4 method to construc Coppe.cosenza objects. It solves de
# TODO(Pessoa) explicar aqui
}
