% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\docType{class}
\name{Factor-class}
\alias{Factor-class}
\title{Factor S4 Class}
\description{
Factor S4 class contains a single slot with the Factor name. A factor in the
COPPE-Cosenza model represents an item to be considered both in the options
and in projects.
}
\section{Slots}{

\describe{
\item{\code{name}}{character}
}}

