% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{Factor}
\alias{Factor}
\title{Factor Constructor}
\usage{
Factor(name)
}
\arguments{
\item{name}{character the factor namecharacter (any other argument will be
cast to character)}
}
\value{
a \code{\link{Factor}} S4 object
}
\description{
Factor(name) is a constructor to Factor S4 objects.
Factor S4 class contains a single slot with the factor name.
}
\examples{
factor <- Factor("name")
Factor("name")


}
