% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors-under-consideration.R
\name{getFactorsUnderConsiderationNames}
\alias{getFactorsUnderConsiderationNames}
\title{Title}
\usage{
getFactorsUnderConsiderationNames(factors.under.consideration)
}
\arguments{
\item{factors.under.consideration}{}
}
\description{
Title
}
