% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-portfolio.R
\name{getOptionPortfolioNames}
\alias{getOptionPortfolioNames}
\title{getOptionPortfolioNames}
\usage{
getOptionPortfolioNames(option.portfolio)
}
\arguments{
\item{option.portfolio}{S4 Option.portfolio object}
}
\value{
vector of character
}
\description{
function that provides a sorted vector with option names.
}
\examples{
\dontrun{getOptionPortfolioNames(option.portfolio)}

}
