% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-portfolio.R
\name{getProjectPortfolioFactors}
\alias{getProjectPortfolioFactors}
\title{Title}
\usage{
getProjectPortfolioFactors(project.portfolio)
}
\arguments{
\item{project.portfolio}{}
}
\description{
Title
}
