% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\docType{class}
\name{Option-class}
\alias{Option-class}
\title{Option S4 Class}
\description{
Option S4 class represents a possible solution to projects. The object
includes a list of Option.resource, which is type checked.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. If other type is provided it will be automatically
cast to character using the \code{\link{as.character()}} function.}

\item{\code{option.resources}}{Option.resources}
}}

