% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{Option}
\alias{Option}
\title{Option Constructor function}
\usage{
Option(name, option.resources)
}
\arguments{
\item{name}{character character (any other argument will be cast to
character)}

\item{option.resources}{Option.resources S4 object. Cannot be empty.}
}
\value{
a \code{\link{Option}} S4 object
}
\description{
Constructs a Option S4 object, which  represents a possible solution to
projects. The object includes a list of Option.resource, which is type
checked.
}
\examples{
\dontrun{Option <- Option(name, option.resources)}



}
