% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{class}
\name{Project-class}
\alias{Project-class}
\title{Project S4 Class}
\description{
Project S4 class represents a potential project and its slots include a
Project.criteria object, with the list of needed factors to the project and
their degree of importance. The project has a non-empty name.
}
\section{Slots}{

\describe{
\item{\code{name}}{character (any other argument will be cast to character)}

\item{\code{project.criteria}}{Project.criteria}
}}

