% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-criteria.R
\name{Project.criteria}
\alias{Project.criteria}
\title{Project.criteria Constructor}
\usage{
Project.criteria(list.of.project.criterion)
}
\arguments{
\item{list.of.project.criterion}{list of Project.criterion S4 objects. The list is type checked
and cannot be empty. The factors of the used project.criterion must be
distinct}
}
\value{
a \code{\link{Project.criteria}} S4 object
}
\description{
Project.criteria is a constructor to Factor S4 objects.
}
\examples{
\dontrun{Project.criteria(list(project.criterion1,project.criterion2))}

}
