% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-portfolio.R
\name{Project.portfolio}
\alias{Project.portfolio}
\title{Project.portfolio Constructor}
\usage{
Project.portfolio(list.of.project)
}
\arguments{
\item{list.of.project}{list of Project S4 objects}
}
\value{
a \code{\link{Project.portfolio}} S4 object
}
\description{
Project.portfolio(list.of.project) is a constructor to Project.portfolio S4 objects.
}
\examples{
\dontrun{Project.portfolio(list(project1, project2, project3))}
# TODO(Taranti) inserir exemplo

}

